Update mapping annotations
Use appropriate annotations in controllers to better correspond
to their purpose.
Use one of the controllers for testing purposes only and clarify
naming of its endpoints.
Move SQL DB blueprint and updated Postman collections to the root
directory of the project because it is of no use to compile them in
resources.